requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
djangorestframework==3.7.3
gunicorn==19.7.1
pytz==2017.3
static3==0.7.0
psycopg2==2.6.1
